pyproject.toml
[project]
name = "mlops-general-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagshub>=0.7.0",
    "dvc[s3]>=3.67.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "mlflow>=2.0",
    "pandas<3",
    "pip>=26.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "ruff>=0.15.11",
    "scikit-learn>=1.8.0",
]
